게시글과 댓글의 수정/삭제 권한 부여하기

+++ app/controllers/foods_controller.rb
@@ -46,6 +46,10 @@ class FoodsController < ApplicationController
 
   def edit
     @post = Post.find(params[:id])
+    if @post.user_id != session[:user_id]
+      flash[:alert] = "수정 권한이 없습니다."
+      redirect_to :back
+    end
   end
 
   def edit_complete
@@ -64,9 +68,14 @@ class FoodsController < ApplicationController
 
   def delete_complete
     post = Post.find(params[:id])
-    post.destroy
-    flash[:alert] = "삭제되었습니다."
-    redirect_to "/"
+    if post.user_id == session[:user_id]
+      post.destroy
+      flash[:alert] = "삭제되었습니다."
+      redirect_to "/"
+    else
+      flash[:alert] = "삭제 권한이 없습니다."
+      redirect_to :back
+    end
   end
   
   def write_comment_complete
@@ -78,6 +87,18 @@ class FoodsController < ApplicationController
 
     flash[:alert] = "새 댓글을 달았습니다."
     redirect_to "/foods/show/#{comment.post_id}"
+  end
+  
+  def delete_comment_complete
+    comment = Comment.find(params[:id])
+    if comment.user_id == session[:user_id]
+      comment.destroy
+      flash[:alert] = "댓글이 삭제되었습니다."
+      redirect_to "/foods/show/#{comment.post_id}"
+    else
+      flash[:alert] = "해당 댓글의 삭제 권한이 없습니다."
+      redirect_to :back
+    end
   end
   
 end
